Jormun: deletion of marshal
Most of the marshal code should be gone with this PR
