+++ source/jormungandr/jormungandr/autocomplete/geocodejson.py
@@ -43,15 +43,6 @@ import requests
 import pybreaker
 from jormungandr import app
 from jormungandr.exceptions import UnknownObject
-from flask_restful import marshal, fields
-from jormungandr.interfaces.v1.fields import (
-    Lit,
-    ListLit,
-    beta_endpoint,
-    feed_publisher_bano,
-    feed_publisher_osm,
-    Integer,
-)
 import re
 
 
@@ -77,14 +68,6 @@ def create_admin_field(geocoding):
             }
         )
     return response
-
-
-class CoordId(fields.Raw):
-    def output(self, key, obj):
-        if not obj:
-            return None
-        lon, lat = get_lon_lat(obj)
-        return '{};{}'.format(lon, lat)
 
 
 def format_zip_code(zip_codes):
@@ -176,195 +159,6 @@ def create_address_field(geocoding, poi_lat=None, poi_lon=None):
     if admins:
         resp['administrative_regions'] = admins
     return resp
-
-
-class AdministrativeRegionField(fields.Raw):
-    """
-    This field is needed to respect Navitia's spec for the sake of compatibility
-    """
-
-    def output(self, key, obj):
-        if not obj:
-            return None
-
-        lon, lat = get_lon_lat(obj)
-        geocoding = obj.get('properties', {}).get('geocoding', {})
-
-        res = {
-            "insee": geocoding.get('citycode') or geocoding.get('city_code'),
-            "level": int(geocoding.get('level')) if geocoding.get('level') else None,
-            "name": geocoding.get('name'),
-            "label": geocoding.get('label'),
-            "id": geocoding.get('id'),
-            "coord": {"lat": lat, "lon": lon},
-            "zip_code": geocoding.get('postcode'),
-        }
-
-        admins = create_administrative_regions_field(geocoding) or create_admin_field(geocoding)
-        if admins:
-            res['administrative_regions'] = admins
-        return res
-
-
-class AddressField(fields.Raw):
-    def output(self, key, obj):
-        if not obj:
-            return None
-
-        lon, lat = get_lon_lat(obj)
-        geocoding = obj.get('properties', {}).get('geocoding', {})
-
-        res = {
-            "id": '{};{}'.format(lon, lat),
-            "coord": {"lon": lon, "lat": lat},
-            "house_number": get_house_number(geocoding.get('housenumber')),
-            "label": geocoding.get('label'),
-            "name": geocoding.get('name'),
-        }
-
-        admins = create_administrative_regions_field(geocoding) or create_admin_field(geocoding)
-        if admins:
-            res['administrative_regions'] = admins
-        return res
-
-
-class PoiField(fields.Raw):
-    def output(self, key, obj):
-        if not obj:
-            return None
-
-        lon, lat = get_lon_lat(obj)
-        geocoding = obj.get('properties', {}).get('geocoding', {})
-        poi_types = geocoding.get('poi_types', [])
-
-        res = {
-            "id": geocoding.get('id'),
-            "coord": {"lon": lon, "lat": lat},
-            "label": geocoding.get('label'),
-            "name": geocoding.get('name'),
-            "properties": {p.get('key'): p.get('value') for p in geocoding.get("properties", [])},
-        }
-        if isinstance(poi_types, list) and poi_types:
-            res['poi_type'] = poi_types[0]
-
-        address = create_address_field(geocoding.get("address"), poi_lat=lat, poi_lon=lon)
-        if address:
-            res['address'] = address
-        admins = create_administrative_regions_field(geocoding) or create_admin_field(geocoding)
-        if admins:
-            res['administrative_regions'] = admins
-        return res
-
-
-class StopAreaField(fields.Raw):
-    def output(self, key, obj):
-        if not obj:
-            return None
-
-        lon, lat = get_lon_lat(obj)
-        geocoding = obj.get('properties', {}).get('geocoding', {})
-
-        resp = {
-            "id": geocoding.get('id'),
-            "coord": {"lon": lon, "lat": lat},
-            "label": geocoding.get('label'),
-            "name": geocoding.get('name'),
-            "timezone": geocoding.get('timezone'),
-        }
-        c_modes = geocoding.get('commercial_modes', [])
-        if c_modes:
-            resp['commercial_modes'] = create_modes_field(c_modes)
-
-        p_modes = geocoding.get('physical_modes', [])
-        if p_modes:
-            resp['physical_modes'] = create_modes_field(p_modes)
-
-        comments = geocoding.get('comments', [])
-        if comments:
-            resp['comments'] = create_comments_field(comments)
-            # To be compatible with old version, we create the "comment" field in addition.
-            # This field is a simple string, so we take only one comment (In our case, the first
-            # element of the list).
-            resp['comment'] = next(iter(comments or []), None).get('name')
-
-        codes = geocoding.get('codes')
-        if codes:
-            resp["codes"] = create_codes_field(codes)
-
-        admins = create_administrative_regions_field(geocoding) or create_admin_field(geocoding)
-        if admins:
-            resp['administrative_regions'] = admins
-
-        prop = {p.get('key'): p.get('value') for p in geocoding.get('properties', [])}
-        if prop:
-            resp['properties'] = prop
-
-        return resp
-
-
-geocode_admin = {
-    "embedded_type": Lit("administrative_region"),
-    "quality": Lit(0),
-    "id": fields.String(attribute='properties.geocoding.id'),
-    "name": fields.String(attribute='properties.geocoding.name'),
-    "administrative_region": AdministrativeRegionField(),
-    "distance": fields.String(attribute='distance', default=None),
-}
-
-
-geocode_addr = {
-    "embedded_type": Lit("address"),
-    "quality": Lit(0),
-    "id": CoordId,
-    "name": fields.String(attribute='properties.geocoding.label'),
-    "address": AddressField(),
-    "distance": fields.String(attribute='distance', default=None),
-}
-
-geocode_poi = {
-    "embedded_type": Lit("poi"),
-    "quality": Lit(0),
-    "id": fields.String(attribute='properties.geocoding.id'),
-    "name": fields.String(attribute='properties.geocoding.label'),
-    "poi": PoiField(),
-    "distance": fields.String(attribute='distance', default=None),
-}
-
-geocode_stop_area = {
-    "embedded_type": Lit("stop_area"),
-    "quality": Lit(0),
-    "id": fields.String(attribute='properties.geocoding.id'),
-    "name": fields.String(attribute='properties.geocoding.label'),
-    "stop_area": StopAreaField(),
-    "distance": fields.String(attribute='distance', default=None),
-}
-
-geocode_feature_func_mapping = {
-    'city': geocode_admin,
-    'administrative_region': geocode_admin,
-    'street': geocode_addr,
-    'house': geocode_addr,
-    'poi': geocode_poi,
-    'public_transport:stop_area': geocode_stop_area,
-}
-
-
-class GeocodejsonFeature(fields.Raw):
-    def format(self, place):
-        type_ = place.get('properties', {}).get('geocoding', {}).get('type')
-
-        if type_ not in geocode_feature_func_mapping:
-            logging.getLogger(__name__).error('Place not serialized (unknown type): {}'.format(place))
-            return None
-
-        return marshal(place, geocode_feature_func_mapping[type_])
-
-
-geocodejson = {
-    "places": fields.List(GeocodejsonFeature, attribute='features'),
-    "warnings": ListLit([fields.Nested(beta_endpoint)]),
-    "feed_publishers": ListLit([fields.Nested(feed_publisher_bano), fields.Nested(feed_publisher_osm)]),
-}
 
 
 class GeocodeJson(AbstractAutocomplete):
