+++ source/jormungandr/jormungandr/interfaces/v1/Coord.py
@@ -28,10 +28,8 @@
 # www.navitia.io
 
 from __future__ import absolute_import, print_function, unicode_literals, division
-from flask_restful import marshal
 from jormungandr import i_manager
 from jormungandr.interfaces.v1.ResourceUri import ResourceUri
-from jormungandr.interfaces.v1.fields import address, context
 from navitiacommon.type_pb2 import _NAVITIATYPE
 import datetime
 from jormungandr.utils import is_coord, get_lon_lat
@@ -39,10 +37,6 @@ import six
 from jormungandr.interfaces.v1.decorators import get_serializer
 from jormungandr.interfaces.v1.serializer import api
 import jormungandr
-from flask_restful.fields import Raw
-
-
-address_marshall_fields = {"regions": Raw, "address": Raw, "message": Raw, "context": context}
 
 
 class Coord(ResourceUri):
@@ -92,7 +86,7 @@ class Coord(ResourceUri):
         )
         return args
 
-    @get_serializer(serpy=api.DictAddressesSerializer, marshall=address_marshall_fields)
+    @get_serializer(serpy=api.DictAddressesSerializer)
     def get(self, region=None, lon=None, lat=None, id=None):
         args = self.parsers["get"].parse_args()
