+++ source/jormungandr/jormungandr/interfaces/v1/GeoStatus.py
@@ -28,34 +28,17 @@
 # www.navitia.io
 
 from __future__ import absolute_import, print_function, unicode_literals, division
-from flask_restful import fields
 from jormungandr import i_manager
 from jormungandr.interfaces.v1.serializer.api import GeoStatusSerializer
 from jormungandr.interfaces.v1.decorators import get_serializer
 from jormungandr.interfaces.v1.StatedResource import StatedResource
-from jormungandr.interfaces.v1.fields import context_utc
-
-geo_status = {
-    'geo_status': fields.Nested(
-        {
-            'street_network_sources': fields.List(fields.String),
-            'nb_admins': fields.Raw,
-            'nb_admins_from_cities': fields.Raw,
-            'nb_ways': fields.Raw,
-            'nb_addresses': fields.Raw,
-            'nb_pois': fields.Raw,
-            'poi_sources': fields.List(fields.String),
-        }
-    ),
-    'context': context_utc,
-}
 
 
 class GeoStatus(StatedResource):
     def __init__(self, *args, **kwargs):
         super(GeoStatus, self).__init__(output_type_serializer=GeoStatusSerializer, *args, **kwargs)
 
-    @get_serializer(serpy=GeoStatusSerializer, marshall=geo_status)
+    @get_serializer(serpy=GeoStatusSerializer)
     def get(self, region=None, lon=None, lat=None):
         region_str = i_manager.get_region(region, lon, lat)
         instance = i_manager.instances[region_str]
