+++ source/jormungandr/jormungandr/interfaces/v1/LineReports.py
@@ -37,37 +37,14 @@ from jormungandr.interfaces.argument import ArgumentDoc
 from jormungandr.interfaces.parsers import default_count_arg_type
 from jormungandr.interfaces.v1.decorators import get_obj_serializer
 from jormungandr.interfaces.v1.errors import ManageError
-from jormungandr.interfaces.v1.fields import (
-    PbField,
-    line,
-    pt_object,
-    NonNullList,
-    NonNullNested,
-    pagination,
-    disruption_marshaller,
-    error,
-    ListLit,
-    beta_endpoint,
-)
 from jormungandr.interfaces.v1.ResourceUri import ResourceUri
 from jormungandr.interfaces.v1.serializer import api
 from jormungandr.resources_utils import ResourceUtc
 from jormungandr.utils import date_to_timestamp
-from flask_restful import fields, reqparse
+from flask_restful import reqparse
 from flask.globals import g
 from datetime import datetime
 import six
-
-
-line_report = {"line": PbField(line), "pt_objects": NonNullList(NonNullNested(pt_object))}
-
-line_reports = {
-    "line_reports": NonNullList(NonNullNested(line_report)),
-    "error": PbField(error, attribute='error'),
-    "pagination": NonNullNested(pagination),
-    "disruptions": fields.List(NonNullNested(disruption_marshaller), attribute="impacts"),
-    "warnings": ListLit([fields.Nested(beta_endpoint)]),
-}
 
 
 class LineReports(ResourceUri, ResourceUtc):
@@ -111,8 +88,7 @@ class LineReports(ResourceUri, ResourceUtc):
             help="If filled, will restrain the search within the given disruption tags",
         )
 
-        self.collection = 'line_reports'
-        self.collections = line_reports
+        self.collection = 'line_reports'  # TODO what?
         self.get_decorators.insert(0, ManageError())
         self.get_decorators.insert(1, get_obj_serializer(self))
