+++ source/jormungandr/jormungandr/interfaces/v1/Places.py
@@ -30,23 +30,13 @@
 # www.navitia.io
 
 from __future__ import absolute_import, print_function, unicode_literals, division
-from flask_restful import fields, abort
+from flask_restful import abort
 from flask.globals import g
 
 from jormungandr.authentication import get_all_available_instances
 from jormungandr.interfaces.v1.decorators import get_serializer
 from jormungandr.interfaces.v1.serializer.api import PlacesSerializer, PlacesNearbySerializer
 from jormungandr import i_manager, timezone, global_autocomplete, authentication
-from jormungandr.interfaces.v1.fields import disruption_marshaller
-from jormungandr.interfaces.v1.fields import (
-    place,
-    NonNullList,
-    NonNullNested,
-    PbField,
-    pagination,
-    error,
-    feed_publisher,
-)
 from jormungandr.interfaces.v1.ResourceUri import ResourceUri
 from jormungandr.interfaces.parsers import default_count_arg_type
 from copy import deepcopy
@@ -68,14 +58,6 @@ from navitiacommon.parser_args_type import (
 )
 from jormungandr.interfaces.common import add_poi_infos_types, handle_poi_infos
 import six
-
-
-places = {
-    "places": NonNullList(NonNullNested(place)),
-    "error": PbField(error, attribute='error'),
-    "disruptions": fields.List(NonNullNested(disruption_marshaller), attribute="impacts"),
-    "feed_publishers": fields.List(NonNullNested(feed_publisher)),
-}
 
 
 class geojson_argument(CustomSchemaType):
@@ -255,16 +237,6 @@ class PlaceUri(ResourceUri):
         return self.api_description(**kwargs)
 
 
-place_nearby = deepcopy(place)
-place_nearby["distance"] = fields.String()
-places_nearby = {
-    "places_nearby": NonNullList(NonNullNested(place_nearby)),
-    "error": PbField(error, attribute='error'),
-    "pagination": PbField(pagination),
-    "disruptions": fields.List(NonNullNested(disruption_marshaller), attribute="impacts"),
-    "feed_publishers": fields.List(NonNullNested(feed_publisher)),
-}
-
 places_types = {
     'stop_areas',
     'stop_points',
@@ -335,7 +307,7 @@ class PlacesNearby(ResourceUri):
         if handle_poi_infos(args["add_poi_infos"], args["bss_stands"]):
             self.get_decorators.insert(1, ManageParkingPlaces(self, 'places_nearby'))
 
-    @get_serializer(serpy=PlacesNearbySerializer, marshall=places_nearby)
+    @get_serializer(serpy=PlacesNearbySerializer)
     def get(self, region=None, lon=None, lat=None, uri=None):
         self.region = i_manager.get_region(region, lon, lat)
         timezone.set_request_timezone(self.region)
