+++ source/jormungandr/jormungandr/interfaces/v1/Status.py
@@ -28,32 +28,19 @@
 # www.navitia.io
 
 from __future__ import absolute_import, print_function, unicode_literals, division
-from flask_restful import fields
 from jormungandr import i_manager, travelers_profile
 from jormungandr.protobuf_to_dict import protobuf_to_dict
-from jormungandr.interfaces.v1.fields import (
-    instance_status_with_parameters,
-    context_utc,
-    ListLit,
-    beta_endpoint,
-    add_common_status,
-)
 from jormungandr.interfaces.v1.serializer.api import StatusSerializer
 from jormungandr.interfaces.v1.decorators import get_serializer
 from jormungandr.interfaces.v1.StatedResource import StatedResource
-
-status = {
-    "status": fields.Nested(instance_status_with_parameters),
-    "context": context_utc,
-    "warnings": ListLit([fields.Nested(beta_endpoint)]),
-}
+from jormungandr.interfaces.v1 import add_common_status
 
 
 class Status(StatedResource):
     def __init__(self, *args, **kwargs):
         super(Status, self).__init__(self, *args, **kwargs)
 
-    @get_serializer(serpy=StatusSerializer, marshall=status)
+    @get_serializer(serpy=StatusSerializer)
     def get(self, region=None, lon=None, lat=None):
         region_str = i_manager.get_region(region, lon, lat)
         response = protobuf_to_dict(
