+++ source/jormungandr/jormungandr/interfaces/v1/Uri.py
@@ -36,7 +36,6 @@ from flask_restful import abort
 from jormungandr.parking_space_availability.parking_places_manager import ManageParkingPlaces
 from jormungandr import i_manager
 from jormungandr.interfaces.v1.converters_collection_type import collections_to_resource_type
-from jormungandr.interfaces.v1.fields import get_collections
 from jormungandr.interfaces.v1.ResourceUri import ResourceUri, protect
 from jormungandr.interfaces.parsers import default_count_arg_type
 from jormungandr.interfaces.v1.errors import ManageError
@@ -219,7 +218,6 @@ def journey_pattern_points(is_collection):
                 "journey_pattern_points",
                 output_type_serializer=api.JourneyPatternPointsSerializer,
             )
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return JourneyPatternPoints
@@ -234,7 +232,6 @@ def commercial_modes(is_collection):
             Uri.__init__(
                 self, is_collection, "commercial_modes", output_type_serializer=api.CommercialModesSerializer
             )
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return CommercialModes
@@ -249,7 +246,6 @@ def journey_patterns(is_collection):
             Uri.__init__(
                 self, is_collection, "journey_patterns", output_type_serializer=api.JourneyPatternsSerializer
             )
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return JourneyPatterns
@@ -264,7 +260,6 @@ def vehicle_journeys(is_collection):
             Uri.__init__(
                 self, is_collection, "vehicle_journeys", output_type_serializer=api.VehicleJourneysSerializer
             )
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return VehicleJourneys
@@ -277,7 +272,6 @@ def trips(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "trips", output_type_serializer=api.TripsSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return Trips
@@ -292,7 +286,6 @@ def physical_modes(is_collection):
             Uri.__init__(
                 self, is_collection, "physical_modes", output_type_serializer=api.PhysicalModesSerializer
             )
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return PhysicalModes
@@ -312,7 +305,6 @@ def stop_points(is_collection):
                 *args,
                 **kwargs
             )
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
             self.parsers["get"].add_argument(
                 "original_id", type=six.text_type, help="original uri of the object you want to query"
@@ -328,7 +320,6 @@ def stop_areas(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "stop_areas", output_type_serializer=api.StopAreasSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
             self.parsers["get"].add_argument(
                 "original_id", type=six.text_type, help="original uri of the object you want to query"
@@ -344,7 +335,6 @@ def connections(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "connections", output_type_serializer=api.ConnectionsSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return Connections
@@ -357,7 +347,6 @@ def companies(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "companies", output_type_serializer=api.CompaniesSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return Companies
@@ -370,7 +359,6 @@ def poi_types(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "poi_types", output_type_serializer=api.PoiTypesSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return PoiTypes
@@ -383,7 +371,6 @@ def routes(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "routes", output_type_serializer=api.RoutesSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
             self.parsers["get"].add_argument(
                 "original_id", type=six.text_type, help="original uri of the object you want to query"
@@ -398,7 +385,6 @@ def line_groups(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "line_groups", output_type_serializer=api.LineGroupsSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
             self.parsers["get"].add_argument(
                 "original_id", type=six.text_type, help="original uri of the object you want to query"
@@ -414,7 +400,6 @@ def lines(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "lines", output_type_serializer=api.LinesSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
             self.parsers["get"].add_argument(
@@ -431,7 +416,6 @@ def pois(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "pois", output_type_serializer=api.PoisSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
             self.parsers["get"].add_argument(
                 "original_id", type=six.text_type, help="original uri of the object you want to query"
@@ -467,7 +451,6 @@ def networks(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "networks", output_type_serializer=api.NetworksSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
             self.parsers["get"].add_argument(
                 "original_id", type=six.text_type, help="original uri of the object you want to query"
@@ -480,7 +463,6 @@ def disruptions(is_collection):
     class Disruptions(Uri):
         def __init__(self):
             Uri.__init__(self, is_collection, "disruptions", output_type_serializer=api.DisruptionsSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
             self.parsers["get"].add_argument(
                 "original_id", type=six.text_type, help="original uri of the object you want to query"
@@ -496,7 +478,6 @@ def contributors(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "contributors", output_type_serializer=api.ContributorsSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return Contributors
@@ -509,7 +490,6 @@ def datasets(is_collection):
 
         def __init__(self):
             Uri.__init__(self, is_collection, "datasets", output_type_serializer=api.DatasetsSerializer)
-            self.collections = get_collections(self.collection)
             self.get_decorators.insert(1, get_obj_serializer(self))
 
     return Datasets
