+++ source/jormungandr/jormungandr/interfaces/v1/__init__.py
@@ -39,3 +39,21 @@ __all__ = [
     'Ptobjects',
     'JSONSchema',
 ]
+
+
+def add_common_status(response, instance):
+    response['status']["is_open_data"] = instance.is_open_data
+    response['status']["is_open_service"] = instance.is_free
+    response['status']['realtime_proxies'] = []
+    for realtime_proxy in instance.realtime_proxy_manager.realtime_proxies.values():
+        response['status']['realtime_proxies'].append(realtime_proxy.status())
+
+    response['status']['street_networks'] = []
+    for sn in instance.street_network_services:
+        response['status']['street_networks'].append(sn.status())
+
+    response['status']['ridesharing_services'] = []
+    for rs in instance.ridesharing_services:
+        response['status']['ridesharing_services'].append(rs.status())
+
+    response['status']['autocomplete'] = instance.autocomplete.status()
