+++ source/jormungandr/jormungandr/interfaces/v1/test/elastic_search_tests.py
@@ -33,7 +33,6 @@ import mock
 from jormungandr.autocomplete.geocodejson import GeocodeJson
 from jormungandr.interfaces.v1 import Places
 from jormungandr.tests.utils_test import MockRequests
-from jormungandr.autocomplete.geocodejson import geocodejson
 from jormungandr.interfaces.v1.serializer.geocode_json import GeocodePlacesSerializer
 from jormungandr.interfaces.v1.decorators import get_serializer
 
@@ -80,7 +79,7 @@ def bragi_house_jaures_feature():
     return house_feature
 
 
-@get_serializer(serpy=GeocodePlacesSerializer, marshall=geocodejson)
+@get_serializer(serpy=GeocodePlacesSerializer)
 def get_response(bragi_response):
     return bragi_response
