+++ source/jormungandr/tests/journey_common_tests.py
@@ -1023,11 +1023,8 @@ class JourneyCommon(object):
         assert len(r['journeys']) == 1
 
     def test_section_fare_zone(self):
-        from jormungandr import app
-
         """
         In a 'stop_point', the section 'fare_zone' should be present if the info is available
-        (only the Serpy serializer has this feature, as Marshall will be deprecated soon)
         """
         r = self.query('/v1/coverage/main_routing_test/stop_points')
         # Only stop point 'stopA' has fare zone info
